fixed bug in exchange calc_exchange_coef (in lsm_driver)
Previously lsm_driver was passing T2m to calc_exchange_coefficient, this was too
sensitive and resulted in instability in the model (plus the z_atm supplied was for
the first model level.  Now passing model level T in, and model appears to be
more stable.  Also implemented the Mahrt (1987) function for stable exchange_C, but that
code crashes, so needs to be further debuged before using.