Added precip distribution factor
Added the local_precip_fraction parameter to mp_parameters namelist.
If local_precip_fraction is less than 1, then only that fraction of precipitation
will be added to the grid cell it is calculated for.  The remaining precipitation
will be spread to the surrounding 8 grid cells using an inverse distance weight.