Added options and function to compute dt more formally.
dt is now computed by default to match the CFL criteria based on finding the grid cell with the maximum
sum of u, v, and w winds associated with it.  This can optionally be set to be multiplied by sqrt(3), and
other stability criteria are available as options, but probably should not be used.  The other criteria
are likely to run faster, but will eventually crash.  If they crash, they could simply be restarted with a
stricter setting for a few days and then relaxed again, this would enable a slightly faster simulation.