Changed handling of lat/lon variables slightly
If u/v lat/lon are not specified in forcing file, they default to the mass
lat/lon variables names.

if u/v lat/lon are not specified in the model file, they are interpolated from
the mass file.

Longitudes are still converted to 0-360 by default, but if they straddle
The 0-longitude line, they are converted to -180 - 180.

Also, some additional info is printed for a debug run including raw geo data