Updates to init, now allows terrain to fall above forcing levels
For pressure level forcing data, the coarse model terrain can fall above the lowest forcing level.
This is now allowed, and for AGL_height, the terrain will be subtracted as is (leading to negative heights)
This should allow vertical interpolation to only select winds that are above the real terrain.