Moved density corrections for linear theory and fixed bug when using the LUT
Density correction (u_hat * 2) is now applied by multipling the linear_mask (=linear_contribution) instead of directly in the LUT
Also, there was a bug in the i,j indexing of the linear_mask that was out of bounds.  This should be fixed now, though
a better fix might be to average in x (for u) and in y (for v) to better account for the grid stagger.
Since the linear_mask is going to be approximate anyway, it shouldn't matter too much at the moment