Detailed linear solution implemented
Previously, the LT LUT was only computed based on the LT solution at a model layer mid-point
For larger layers and high stability, this can lead to aliasing between the LT wave pattern
and the model layers.  This solutions permits a minimum layer thickness, and for model layers
thicker than that size, the LUT will be computing by breaking that layer into N layers,
each of which is equal to or smaller than the minimum thickness.

Minimum layer thickness is initially hardcoded as 100 m, but should be made into a parameter.