BUGFIX in detailed winds calculation
previously, each new layer was starting at 0 m AGL and incrementing by step size instead of starting
at the bottom of the actual layer it was being computed for.
Also, BUGFIX in the parallelization: min_layer_size was listed as private, not listed as firstprivate
so it's value will be copied into the parallel region.