updating geo interpolation to fix a subtle bug
issue #35 geointerpolation now check that a given point is inside the polygon.
this could be combined with the previous method (which worked 99.9+% of the time) so that the older (faster)
version is used when possible, but when it fails the new method figure out the correct points to interpolate from.

This commit compiles but is untested