geo-interpolation added idw and small bugfix
fixed bug in point orders for bilinear interpolation
note that this is still not really a perfect interpolation, there will be artifacts on rotated grids
added an inverse distance method instead.  This also produces some artifacts, but maybe less severe?
also, netCDF io debug code is present in this commit