Relaxed constraints on geo_reader point_in_poly tests
This avoids being so picky that points which fall nearly on a line may not count as 'in' the polygon
on either side of the line due to rounding/precision errors.  To make this work the constraints on polygon weights are now
relaxed slightly as well.  At current settings, a point is required to fall within ~1m of the edge of the grid cell to
count as 'in' the grid cell