minor bug fixes in u/v input
Previously it was passing info from the u_grid into a routine processing the v_grid
Probably this was always the same, but it was a problem.
Also, previously the v_grid was still recalculating window size based on the input grid.
This is now removed so it will use the value from the options file.