Fixed initial bugs in caf linear winds
Now compiles and computes LUT, then writes LUT to disk (in a very crude manner)
Need to check that the LUT is computed for the correct region (e.g. how is the stagger accounted for?)
Need to modify spatial winds to use the correct portion of the LUT when pulling LUT data.
Need to use the lut_io capabilities to read and write the LUTs better.