Fixed minor (potential) issue in linear winds
Previously, the temporary_u/v variables were still on the mass grid, even though they were restaggered
While this stored all the data, it meant that the transfer to remote images would fall off the end of the array
for the top/right edge images.  Now the extended data are stored in the staggered grid, note buffer must be >=1 internally.