Made io_read_6d expect a coarray instead of a regular array
Not sure if this is an issue with ifort or actually required by the fortran spec. io_read_6d gets
passed a coarray but only the local copy of it, so it shouldn't matter (and doesn't to gfortran)
However, if io_read_6d allocates the data, I don't know what gfortran does when it doesn't see a coarray allocation.
It doesn't matter in this case because if we are using the read, it doesn't need to be a coarray... still seems broken in gfortran.