Update precision of geo_reader point on line test again.
Now only requires a point to be within 10 m of the border between two grid cells to count as being in either grid cell.
Eventually it might be better to move to double precision for some of these values to permit greater accuracy,
but the current accuracy is still 100x finer than most model grid spacings so it is not a high priority.