Adds a new wind option!
If wind=2 in the physics options and space_varying = true in z_levels, ICAR will now use the
compression of the z coordinate due to topography to accelerate (and decelerate) winds by the
same ratio, this minimizes lifting without leading to the overly enhanced mixing or issues
due to spatial variability in the linear wind field.  However, it also doesn't let topography
affect winds more physically.  This is not as good an option in theory, but may be a better option
in practice.